Perform service restarts with runit rather than ctr
Doing this with ctr is technically possible, but between it and the
linuxkit service command there are interactions that are poorly
documented at best, and completely broken most of the rest of the
time.  The solution instead was to just import runit and then reload
only what was necessary without restarting the outer containerd shim.
Overall this appears to be more stable, and it looks like its mostly
working.